Add new voting tags query to return only used tags (#3240)
